requirements.txt
# Production dependencies

# Translation
langdetect>=1.0.9
deep-translator>=1.11.4

# Text Extraction (OCR)
pytesseract>=0.3.13
Pillow>=10.0.0
pdf2image>=1.17.0
# PyMuPDF>=1.23.0  # For future digital PDF extraction

# Azure Document Intelligence OCR (formerly Form Recognizer)
azure-ai-documentintelligence>=1.0.0

# Interactive CLI
questionary>=2.0.0

# Environment variables
python-dotenv>=1.0.0
